Moved unit tests for GetDeckNotes() into ClassRepositoryTests.
