+++ AnkiJapaneseFlashcardManagerTest/Anki2CardControllerTests.cs
@@ -16,57 +16,6 @@ namespace AnkiJapaneseFlashcardManagerTests
 		string _anki2FolderPath = "./Resources/Anki2 Files/";
 
 		[Theory]
-		[InlineData("empty_random_decks.anki2", 1706982318565)]
-		[InlineData("empty_random_decks.anki2", 1706982351536)]
-		[InlineData("empty_kanjiResource_newKanji_decks.anki2", 1706982318565)]
-		[InlineData("empty_kanjiResource_newKanji_decks.anki2", 1706982351536)]
-		[InlineData("empty_kanjiResource_newKanji_decks.anki2", 1707160682667)]
-		[InlineData("empty_kanjiResource_newKanji_decks.anki2", 1707160947123)]
-		public void No_notes_found_in_a_deck_is_empty(string anki2File, long deckId)
-		{
-			//Arrange
-			Anki2Context dbContext = new Anki2Context(_anki2FolderPath + anki2File);
-			Anki2Controller anki2Controller = new Anki2Controller(dbContext);
-
-			//Act
-			var notes = anki2Controller.GetDeckNotes(deckId);
-
-			//Assert
-			notes.Should().BeEmpty();
-		}
-
-		[Theory]
-		[InlineData("飲newKanji_食欠人良resourceKanji_decks.anki2", 1707160947123)]
-		[InlineData("飲newKanji_食欠人良resourceKanji_decks.anki2", 1707160682667)]
-		public void Get_notes_by_deck_id(string anki2File, long deckId)//TODO: Refactor to check for expected values
-		{
-			//Arrange
-			Anki2Context dbContext = new Anki2Context(_anki2FolderPath + anki2File);
-			Anki2Controller anki2Controller = new Anki2Controller(dbContext);
-
-			//Act
-			var notes = anki2Controller.GetDeckNotes(deckId);
-
-			//Assert
-			notes.Should().NotBeEmpty();
-		}
-
-		[Theory]
-		[InlineData("deck_with_different_card_types.anki2", 1707263514556, new[] { 1707263555296, 1707263973429, 1707263567670 })]
-		public void Mulitple_card_entries_of_the_same_note_is_a_distinct_list_of_notes(string anki2File, long deckId, long[] expectedNoteIds)
-		{
-			//Arrange
-			Anki2Context dbContext = new Anki2Context(_anki2FolderPath + anki2File);
-			Anki2Controller anki2Controller = new Anki2Controller(dbContext);
-
-			//Act
-			var notes = anki2Controller.GetDeckNotes(deckId);
-
-			//Assert
-			notes.Select(n => n.Id).Should().BeEquivalentTo(expectedNoteIds);
-		}
-
-		[Theory]
 		[InlineData("飲newKanji_食欠人良resourceKanji_decks.anki2", new[] { 1707169497960, 1707169570657, 1707169983389, 1707170000793 }, 1707160682667)]
 		public void Move_notes_between_decks(string anki2File, long[] noteIdsToMove, long deckIdToMoveTo)
 		{
